Human - replay stores mmr as a long for some reason, so consistently match that

+++ StatCraft/StatCraft/Models/GameData/GamePlayer.cs
@@ -8,7 +8,7 @@ namespace StatCraft.Models.GameData
     {
         public required string Clan { get; set; }
         public required string Name { get; set; }
-        public required int Mmr { get; set; }
+        public required long Mmr { get; set; }
         public required char Race { get; set; }
         public required bool Random { get; set; }
     }
